# OPT: Speed up _check_git_annex_version()