--- a/datalad/support/annexrepo.py
+++ b/datalad/support/annexrepo.py
@@ -47,7 +47,6 @@ from datalad.utils import (
     assure_list,
     auto_repr,
     ensure_list,
-    get_linux_distribution,
     on_windows,
     partition,
     Path,
@@ -555,15 +554,8 @@ class AnnexRepo(GitRepo, RepoInterface):
     def _check_git_annex_version(cls):
         ver = external_versions['cmd:annex']
         # in case it is missing
-        msg = "Visit http://git-annex.branchable.com/install/"
-        # we might be able to do better
-        try:
-            linux_distribution_name = get_linux_distribution()[0]
-            if linux_distribution_name in {'debian', 'ubuntu'}:
-                msg = "Install  git-annex-standalone  from NeuroDebian " \
-                      "(http://neuro.debian.net)"
-        except:  # pragma: no cover
-            pass
+        msg = "Visit http://handbook.datalad.org/r.html?install " \
+              "for instructions on how to install DataLad and git-annex."
 
         exc_kwargs = dict(
             name="git-annex",