It removes a programmatic assesment of the platform in order to
build a targeted message to be displayed whenever git-annex
cannot be found.

Now the message is generic and saved 1.5s for a wide range of
datalad command executed on the cmdline.

For example, a `datalad create` on Debian drops from 2.5s to 800ms.

now:
```
(datalad3-dev) mih@meiner /tmp % time datalad create s1
[INFO   ] Creating a new annex repo at /tmp/s1 
[INFO   ] Scanning for unlocked files (this may take some time) 
create(ok): /tmp/s1 (dataset)
datalad create s1  0.59s user 0.27s system 104% cpu 0.833 total

(datalad3-dev) mih@meiner /tmp/s1 (git)-[master] % touch this
(datalad3-dev) mih@meiner /tmp/s1 (git)-[master] % time datalad save
add(ok): this (file)                                                                                                                       
save(ok): . (dataset)                                                                                                                      
action summary:                                                                                                                            
  add (ok: 1)
  save (ok: 1)
datalad save  0.50s user 0.17s system 98% cpu 0.677 total

(datalad3-dev) mih@meiner /tmp/s1 (git)-[master] % time datalad status --annex
1 annex'd file (0.0 B recorded total size)
nothing to save, working tree clean
datalad status --annex  0.32s user 0.09s system 102% cpu 0.403 total
```

before:
```
(datalad3-dev) mih@meiner /tmp % time datalad create s2
[INFO   ] Creating a new annex repo at /tmp/s2 
[INFO   ] Scanning for unlocked files (this may take some time) 
create(ok): /tmp/s2 (dataset)
datalad create s2  2.02s user 0.34s system 98% cpu 2.413 total

(datalad3-dev) mih@meiner /tmp/s2 (git)-[master] % time datalad save
add(ok): this (file)                                                                                                                       
save(ok): . (dataset)                                                                                                                      
action summary:                                                                                                                            
  add (ok: 1)
  save (ok: 1)
datalad save  2.04s user 0.31s system 92% cpu 2.531 total

(datalad3-dev) mih@meiner /tmp/s2 (git)-[master] % time datalad status --annex
1 annex'd file (0.0 B recorded total size)
nothing to save, working tree clean
datalad status --annex  1.88s user 0.20s system 100% cpu 2.059 total
```


Fixes #5317