@kyleam 2021-01-12
Nice.

I'm guessing this speed is mainly for py3.8 and above where `distro` is used.  On my buster system with Python 3.7.3:

```console
# master (4ac10eb04b)
$ multitime -n 5 datalad status --annex             
nothing to save, working tree clean                         
nothing to save, working tree clean
nothing to save, working tree clean
nothing to save, working tree clean
nothing to save, working tree clean
===> multitime results
1: datalad status --annex
            Mean        Std.Dev.    Min         Median      Max
real        0.719       0.011       0.705       0.716       0.735       
user        0.589       0.023       0.550       0.590       0.622       
sys         0.127       0.016       0.109       0.121       0.153       

# pr (af8174123d)
$ multitime -n 5 datalad status --annex
nothing to save, working tree clean
nothing to save, working tree clean
nothing to save, working tree clean
nothing to save, working tree clean
nothing to save, working tree clean
===> multitime results
1: datalad status --annex
            Mean        Std.Dev.    Min         Median      Max
real        0.711       0.017       0.680       0.715       0.731       
user        0.591       0.024       0.547       0.593       0.612       
sys         0.116       0.012       0.101       0.116       0.130
```