@mih 2021-01-12
I can confirm that I see no speed-up on a different system (buster, 3.7.3)