@yarikoptic 2021-01-12
<details>
<summary>FWIW, on my laptop - total 0.520 cold .380 hot</summary>

```shell
$> time python3 -c 'import distro; print(distro.linux_distribution())'
('Debian GNU/Linux', 'testing', 'bullseye')
python3 -c 'import distro; print(distro.linux_distribution())'  0.34s user 0.14s system 91% cpu 0.520 total

$> time python3 -c 'import distro; print(distro.linux_distribution())'
('Debian GNU/Linux', 'testing', 'bullseye')
python3 -c 'import distro; print(distro.linux_distribution())'  0.28s user 0.10s system 100% cpu 0.380 total
```
</details>